refactor: ErrorCategory / LogRecord をドメイン層からアプリケーション層の port モジュールへ移動
ロギングは横断的関心事であり、ドメイン層が持つべき概念ではない。
application/port.rs を新設し ErrorCategory・LogRecord・ErrorLogger を集約することで、
インフラ層が application::port を介してポートアダプタパターンに従って実装できるようにする。
