+++ src/contexts/evaluation/application.rs
@@ -1,5 +1,6 @@
 pub mod config_service;
 pub mod errors;
+pub mod port;
 pub mod prompt;
 
 use crate::contexts::evaluation::domain::pr_state::NotApplicableState;
